move to ESM
+ update chalk
